requirements.txt
pytest
pytest-selenium
allure-pytest
pyyaml
pytest-check
colorama
Appium-Python-client
